requirements.txt
yfinance
streamlit
plotly.graph_objects
plotly.express 
plotly.subplots 
pandas_ta as ta
datetime
talib 
figure